gyp: Fix creating dependency files on Dos-like systems.
On systems with drive letters the generated file would attempt to create
and use fileas like D:/foo/bar/D:/foo/bar/abc.d which would eventually
fail because of ':' in the middle. Removing it fixes the problem (and
should be platform-agnostic).

Interesting that creating directories like that with GNU 'mkdir' succeeds
on OS/2 (with D: completely removed from the created path) which may
be a special mkdir hack in order to overcome this typical problem of
concatenating two paths which works fine on Unix-like systems but
presents various sorts of problems on Dos-like ones.